fix(console): make cross-arch client bundles deterministic for real
bun silently ignores NODE_OPTIONS=--require, so the sorted-readdir
shim never ran inside the bun-based image builds: route readdir order
came from each runner's overlayfs, the ARM and Intel builds emitted
different bundle hashes, and any HTML answered by one arch's pod
404'd its immutable assets on the other arch's pod (dead hydration
in production).

Load the shim as an ESM side-effect import at the top of each app's
vite.config.ts instead — kit reads directories via the default
node:fs import, so patching the module object intercepts every call
under Node and bun alike. Sort all src/ reads (routes drive node IDs,
params drive the matcher list). A new CI step in the manifest job
diffs the client asset lists of both arch images and fails the
publish if they ever diverge again.
